# Implement layout="zip" for Lambda/GCF, skipping lambdex